+++ src/main/java/botcommons/commands/AutoCompleteHandler.java
@@ -1,0 +1,9 @@
+package botcommons.commands;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AutoCompleteHandler {
+	String[] value();
+}
