+++ src/main/java/botcommons/commands/CommandHolder.java
@@ -1,0 +1,9 @@
+package botcommons.commands;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CommandHolder {
+	String group() default "N/A";
+}
