+++ src/main/java/botcommons/menu/Menu.java
@@ -1,0 +1,9 @@
+package botcommons.menu;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Menu {
+	String id();
+}
