+++ src/main/java/botcommons/menu/MenuManager.java
@@ -1,0 +1,117 @@
+package botcommons.menu;
+
+import net.dv8tion.jda.api.JDA;
+import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
+import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+import net.dv8tion.jda.api.interactions.InteractionHook;
+import net.dv8tion.jda.api.interactions.components.ActionRow;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import net.dv8tion.jda.api.requests.restaction.MessageCreateAction;
+import net.dv8tion.jda.api.requests.restaction.WebhookMessageCreateAction;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+@SuppressWarnings("unused")
+public class MenuManager extends ListenerAdapter {
+	private final JDA jda;
+	private static final Map<String, IMenu> menuRegistry = new HashMap<>();
+
+	public static MenuManager instance;
+
+	public MenuManager(JDA jda) {
+		instance = this;
+		this.jda = jda;
+	}
+
+	public static void registerMenu(IMenu... menus) {
+		for (IMenu menu : menus) {
+			String menuId = menu.getClass().getAnnotation(Menu.class).id();
+
+			menuRegistry.put(menuId, menu);
+		}
+	}
+
+	public static void sendMenu(String menuId, String channelId) {
+		IMenu menu = menuRegistry.get(menuId);
+
+		if (menu==null)
+			throw new IllegalArgumentException("Menu not found " + menuId);
+
+		TextChannel channel = instance.jda.getTextChannelById(channelId);
+
+		if (channel==null) {
+			throw new IllegalArgumentException("Channel not found " + channelId);
+		}
+
+		channel.sendMessageEmbeds(menu.build()).addActionRow(menu.getButtons())
+				.queue(message -> menu.setMessageId(message.getId()));
+	}
+
+	public static void sendMenuPrivate(String menuId, String userId) {
+		IMenu menu = menuRegistry.get(menuId);
+
+		if (menu == null)
+			throw new IllegalArgumentException("Menu not found " + menuId);
+
+		instance.jda.retrieveUserById(userId).queue(user ->
+				user.openPrivateChannel().queue(channel ->
+						channel.sendMessageEmbeds(menu.build()).addActionRow(menu.getButtons()).queue(message ->
+								menu.setMessageId(message.getId()))));
+	}
+
+	private static IMenu getMenu(String menuId, Object... args) {
+		IMenu menu = menuRegistry.get(menuId);
+		try {
+			if (menu.getClass().getConstructors().length>1&&args.length>0)
+				menu = (IMenu) Arrays.stream(menu.getClass().getConstructors()).filter(constructor ->
+						constructor.getParameterCount()!=0).toList().getFirst().newInstance(args);
+			if (menuRegistry.containsKey(menuId))
+				menuRegistry.replace(menuId, menu);
+			else
+				registerMenu(menu);
+		} catch (Exception ignored) {}
+		if (menu == null)
+			menu = menuRegistry.get(menuId);
+		if (menu == null)
+			throw new IllegalArgumentException("Menu not found " + menuId);
+		return menu;
+	}
+
+
+	public static void replyMenu(String menuId, Message message, Object... args) {
+		IMenu menu = getMenu(menuId, args);
+
+		splitButtons(message.replyEmbeds(menu.build()), menu.getButtons()).queue(message1 ->
+				menu.setMessageId(message1.getId()));
+	}
+
+	public static void replyMenu(String menuId, InteractionHook hook, Object... args) {
+		IMenu menu = getMenu(menuId, args);
+
+		splitButtons(hook.sendMessageEmbeds(menu.build()), menu.getButtons()).queue(message1 ->
+				menu.setMessageId(message1.getId()));
+	}
+
+	private static MessageCreateAction splitButtons(MessageCreateAction action, Button... buttons) {
+		ActionRow.partitionOf(buttons).forEach(row -> action.addActionRow(row.getButtons()));
+		return action;
+	}
+
+	private static WebhookMessageCreateAction<Message> splitButtons(WebhookMessageCreateAction<Message> action, Button... buttons) {
+		ActionRow.partitionOf(buttons).forEach(row -> action.addActionRow(row.getButtons()));
+		return action;
+	}
+
+
+	@Override
+	public void onButtonInteraction(ButtonInteractionEvent event) {
+		String messageId = event.getMessageId();
+		menuRegistry.values().stream()
+				.filter(m -> m.getMessageId() != null && m.getMessageId().equals(messageId))
+				.findFirst().ifPresent(menu -> menu.handle(event));
+	}
+}
