--- src/main/java/dev/scyye/botcommons/commands/AutoCompleteHandler.java
@@ -1,9 +1,0 @@
-package dev.scyye.botcommons.commands;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AutoCompleteHandler {
-	String[] value();
-}
