--- src/main/java/dev/scyye/botcommons/commands/CommandHolder.java
@@ -1,9 +1,0 @@
-package dev.scyye.botcommons.commands;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CommandHolder {
-	String group() default "N/A";
-}
