--- src/main/java/dev/scyye/botcommons/menu/Menu.java
@@ -1,9 +1,0 @@
-package dev.scyye.botcommons.menu;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Menu {
-	String id();
-}
